feat: mpa default tabbar (#4323)
* feat: mpa default tabbar

* chore: change pageName

* chore: remove useless change

* chore: sync tabbar path

* chore: remove generate rax template file

* chore: tab bar component order
