--- packages/plugin-app-core/src/generator/templates/app/rax/ErrorBoundary/ErrorBoundaryFallback.tsx.ejs
@@ -1,47 +1,0 @@
-import { createElement, CSSProperties } from 'rax';
-
-interface IProps {
-  componentStack: string;
-  error: Error;
-}
-
-const toTitle = (error: Error): string => {
-  return `${error.toString()}`;
-};
-
-const style = {
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  margin: '100px 0',
-  color: '#ed3131'
-} as CSSProperties;
-
-const ErrorBoundaryFallback = ({ error }: IProps) => {
-  return (
-    <div style={style} title={toTitle(error)}>
-      <svg
-        viewBox="0 0 1024 1024"
-        version="1.1"
-        xmlns="http://www.w3.org/2000/svg"
-        p-id="843"
-        width="60"
-        height="60"
-      >
-        <path
-          d="M1024 512C1024 229.23 794.77 0 512 0S0 229.23 0 512s229.23 512 512 512c117.41 0 228.826-39.669 318.768-111.313 10.79-8.595 12.569-24.308 3.975-35.097-8.594-10.789-24.308-12.568-35.097-3.974C718.47 938.277 618.002 974.049 512 974.049 256.818 974.049 49.951 767.182 49.951 512S256.818 49.951 512 49.951 974.049 256.818 974.049 512c0 87.493-24.334 171.337-69.578 243.96-7.294 11.708-3.716 27.112 7.992 34.405 11.707 7.294 27.11 3.716 34.405-7.991C997.014 701.88 1024 608.898 1024 512z"
-          p-id="844"
-          fill="#cdcdcd"
-        />
-        <path
-          d="M337.17 499.512c34.485 0 62.44-27.955 62.44-62.439s-27.955-62.439-62.44-62.439c-34.483 0-62.438 27.955-62.438 62.44 0 34.483 27.955 62.438 62.439 62.438z m374.635 0c34.484 0 62.439-27.955 62.439-62.439s-27.955-62.439-62.44-62.439c-34.483 0-62.438 27.955-62.438 62.44 0 34.483 27.955 62.438 62.439 62.438zM352.788 704.785c43.377-34.702 100.364-55.425 171.7-55.425 71.336 0 128.322 20.723 171.7 55.425 26.513 21.21 42.695 42.786 50.444 58.284 6.168 12.337 1.168 27.34-11.17 33.508-12.337 6.169-27.34 1.168-33.508-11.17-0.918-1.834-3.462-6.024-7.788-11.793-7.564-10.084-17.239-20.269-29.183-29.824-34.671-27.737-80.71-44.478-140.495-44.478-59.786 0-105.824 16.74-140.496 44.478-11.944 9.555-21.619 19.74-29.182 29.824-4.327 5.769-6.87 9.959-7.788 11.794-6.169 12.337-21.171 17.338-33.509 11.17-12.337-6.17-17.338-21.172-11.169-33.509 7.75-15.498 23.931-37.074 50.444-58.284z"
-          p-id="845"
-          fill="#cdcdcd"
-        />
-      </svg>
-      <h3>Oops! Something went wrong.</h3>
-    </div>
-  );
-};
-
-export default ErrorBoundaryFallback;
